partially revert 6c4614f13f6b5b009361e5b4b833e7db986c2ee9 to fix sound issue

+++ include/vlc_atomic.h
@@ -471,7 +471,7 @@ static inline void vlc_atomic_store_float(vlc_atomic_float *atom, float f)
 {
     union { float f; uint32_t i; } u;
     u.f = f;
-    atomic_store(atom, (short) u.i);
+    atomic_store(atom, u.i);
 }
 
 # else /* C++ */
